Safari-609.1.11.1/LayoutTests/plugins/reloadplugins-no-pages-expected.txt

ALERT: Page was loaded

Test for NPN_RelaodPlugin in which reloadOpenPages is false. You should see one alert, indicating that the page was loaded once.

SUCCESS
